server/requirements.txt
openenv-core[core]>=0.2.2
openai>=1.30.0
pydantic>=2.7.0
uvicorn[standard]>=0.30.0
